Swap the recovery commands (/cache/recovery/command) around
that it first flashes the zip and only afterwards formats the cache, as the update file resides on the /cache partition, so the update file can't be found after formatting the cache partition. For some weird reason this behaviour doesn't cause any issue on FP Open TWRP & the stock recovery.

Change-Id: I2946b2ac9f493102b0139e8b1ea175794765952d
Signed-off-by: Luca Weiss <luca@z3ntu.xyz>